chore: rename actual files, update storybook docs further
