+++ src/core/components/avatars/avatarBase/avatarBase.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, type ComponentPropsWithoutRef } from 'react';
-import { useGukCoreContext } from '../../odsCoreProvider';
+import { useGukCoreContext } from '../../gukCoreProvider';
 
 export interface IAvatarBaseProps extends ComponentPropsWithoutRef<'img'> {}
