+++ src/core/components/dataList/dataListFilter/dataListFilter.tsx
@@ -2,8 +2,8 @@ import classNames from 'classnames';
 import { useState, type ChangeEvent, type ComponentProps } from 'react';
 import { AvatarIcon } from '../../avatars';
 import { Button } from '../../button';
+import { useGukCoreContext } from '../../gukCoreProvider';
 import { Icon, IconType } from '../../icon';
-import { useGukCoreContext } from '../../odsCoreProvider';
 import { Spinner } from '../../spinner';
 import { useDataListContext } from '../dataListContext';
 import { DataListFilterSort } from './dataListFilterSort';
