+++ src/core/components/gukCoreProvider/gukCoreProvider.mdx
@@ -1,5 +1,5 @@
 import { ArgTypes, Meta, Description, Title, Subtitle } from '@storybook/blocks';
-import { GukCoreProvider } from './odsCoreProvider';
+import { GukCoreProvider } from './gukCoreProvider';
 
 <Meta title="Core/Components/GukCoreProvider" />
 
@@ -14,7 +14,7 @@ the `<Image />` and `<Link />` components from Next.js, or when specific copy te
 **Usage example**
 
 ```typescript
-import { GukCoreProvider } from '@aragon-guk';
+import { GukCoreProvider } from '@aragon/gov-ui-kit';
 import NextImage from 'next/image';
 import NextLink from 'next/link';
