+++ src/core/components/index.ts
@@ -11,11 +11,11 @@ export * from './dialogs';
 export * from './documentParser';
 export * from './dropdown';
 export * from './forms';
+export * from './gukCoreProvider';
 export * from './heading';
 export * from './icon';
 export * from './illustrations';
 export * from './link';
-export * from './odsCoreProvider';
 export * from './progress';
 export * from './rerender';
 export * from './spinner';
