+++ src/modules/components/address/addressInput/addressInput.tsx
@@ -18,8 +18,8 @@ import {
 import { ChainEntityType, useBlockExplorer } from '../../../hooks';
 import type { IWeb3ComponentProps } from '../../../types';
 import { addressUtils, ensUtils } from '../../../utils';
+import { useGukModulesContext } from '../../gukModulesProvider';
 import { MemberAvatar } from '../../member';
-import { useGukModulesContext } from '../../odsModulesProvider';
 
 export interface IAddressInputResolvedValue {
     /**
