+++ src/modules/components/gukModulesProvider/gukModulesProvider.mdx
@@ -1,5 +1,5 @@
 import { ArgTypes, Meta, Description, Title, Subtitle } from '@storybook/blocks';
-import { GukModulesProvider } from './odsModulesProvider';
+import { GukModulesProvider } from './gukModulesProvider';
 
 <Meta title="Modules/Components/GukModulesProvider" />
