+++ src/modules/components/index.ts
@@ -1,8 +1,8 @@
 export * from './address';
 export * from './asset';
 export * from './dao';
+export * from './gukModulesProvider';
 export * from './member';
-export * from './odsModulesProvider';
 export * from './proposal';
 export * from './transaction';
 export * from './vote';
