+++ src/modules/components/proposal/proposalActions/proposalActionsAction/proposalActionsActionDecodedView/proposalActionsActionDecodedView.tsx
@@ -1,5 +1,5 @@
 import { InputText } from '../../../../../../core';
-import { useGukModulesContext } from '../../../../odsModulesProvider';
+import { useGukModulesContext } from '../../../../gukModulesProvider';
 import type { IProposalAction } from '../../proposalActionsTypes';
 
 export interface IProposalActionsActionDecodedViewProps {
