+++ src/modules/components/wallet/wallet.test.tsx
@@ -2,7 +2,7 @@ import { type QueryClient } from '@tanstack/react-query';
 import { render, screen } from '@testing-library/react';
 import { sepolia } from 'viem/chains';
 import * as wagmi from 'wagmi';
-import { GukModulesProvider } from '../odsModulesProvider';
+import { GukModulesProvider } from '../gukModulesProvider';
 import { Wallet, type IWalletProps } from './wallet';
 
 jest.mock('../member', () => ({
