+++ src/modules/index.mdx
@@ -37,5 +37,5 @@ In order to use the GovKit Modules components, make sure to:
     }
     ```
 
-    Refer to the [GukModulesProvider doc](/docs/modules-components-odsmodulesprovider--docs) for further details about
+    Refer to the [GukModulesProvider doc](/docs/modules-components-gukmodulesprovider--docs) for further details about
     the provider.
